MaxMin исправление: 1. Вместо first, second и др. в именах лучше использовать цифры
2. По заданию здесь должны быть целые числа.
Но этот пункт можно не исправлять

3. Нужно также найти минимум через тернарный оператор

+++ src/main/java/AcademItSchool/LessonThree/MaxMin.java
@@ -20,8 +20,8 @@ public class MaxMin {
             System.out.printf("Наибольшее число = %.2f%nНаименьшее число = %.2f%n", numberTwo, numberOne);
         }
 
-        double max = (numberOne > numberTwo) ? numberOne : numberTwo;
-        double min = (numberOne < numberTwo) ? numberOne : numberTwo;
+        int max = (numberOne > numberTwo) ? numberOne : numberTwo;
+        int min = (numberOne < numberTwo) ? numberOne : numberTwo;
         System.out.printf("Наибольшее число = %.2f%nНаименьшее число = %.2f%n", max, min);
     }
 }
